Move flask_common.utils.legacy.custom_query_counter to flask_common.mongo.query_counters.custom_query_counter
